+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#52](https://github.com/rubocop-hq/rubocop-rails/issues/52): Add new `Rails/AfterCommitOverride` cop. ([@fatkodima][])
 * [#274](https://github.com/rubocop-hq/rubocop-rails/pull/274): Add new `Rails/WhereNot` cop. ([@fatkodima][])
 * [#311](https://github.com/rubocop-hq/rubocop-rails/issues/311): Make `Rails/HelperInstanceVariable` aware of memoization. ([@koic][])
+* [#330](https://github.com/rubocop-hq/rubocop-rails/pull/330): Add new `Rails/ExistsById` cop. ([@tabakazu][])
 
 ### Bug fixes
 
@@ -268,3 +269,4 @@
 [@philcoggins]: https://github.com/philcoggins
 [@kunitoo]: https://github.com/kunitoo
 [@jaredmoody]: https://github.com/jaredmoody
+[@tabakazu]: https://github.com/tabakazu